# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "tribd",
    "description": "Tributary daemon API",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1.1.1"
  },
  "paths": {
...
    "/api/v1/takes/{take}": {
      "delete": {
        "summary": "Permanently remove one take — audio, peaks and manifest together.",
        "description": "Refused while recording, and refused for the take being played: the\nfeeder threads hold open readers and reopen the files on every loop\npass, so unlinking underneath them would run playback off unlinked\ninodes until a later wrap died somewhere confusing.",
        "operationId": "delete_take",
        "parameters": [
          {
            "name": "take",
            "in": "path",
            "description": "Take number",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted; the remaining takes, newest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TakeDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No such take"
          },
          "409": {
            "description": "Recording, or that take is playing"
          }
        }
      }
    },
    "/api/v1/takes/{take}/peaks": {
      "get": {
        "summary": "Waveform peaks for every track of a take, in `TakeDto.tracks` order.",
        "description": "Binary layout (little-endian): `\"TPKS\"` · `version: u8 = 1` · `pad: u8`\n· `track_count: u16` · `samples_per_bin: u32` · then per track:\n`bin_count: u32` + `bin_count × (min: i16, max: i16)`.",
        "operationId": "take_peaks",
        "parameters": [
          {
            "name": "take",
            "in": "path",
            "description": "Take number",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Binary peaks document",
            "content": {
              "application/octet-stream": {}
            }
          },
          "404": {
            "description": "No such take"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TakeDto": {
        "type": "object",
        "required": [
          "take",
          "started_at_unix",
          "sample_rate",
          "damaged",
          "duration_secs",
          "tracks"
        ],
        "properties": {
          "damaged": {
            "type": "boolean"
          },
          "duration_secs": {
            "type": "number",
            "format": "double"
          },
          "sample_rate": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "started_at_unix": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "take": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "tracks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TakeTrackDto"
            }
          }
        }
      },
      "TakeTrackDto": {
        "type": "object",
        "required": [
          "file",
          "channels",
          "frames",
          "dropped_samples"
        ],
        "properties": {
          "channels": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "dropped_samples": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "file": {
            "type": "string"
          },
          "frames": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "strip_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "The strip this track tapped — absent for the master mix and for\ntakes cut before the field existed.",
            "minimum": 0
          }
        }
      },
...
    }
  }
}